backend/requirements.txt
fastapi
uvicorn
google-generativeai
neo4j
pydantic-settings
python-jose[cryptography]
httpx
python-multipart
python-dotenv
apscheduler
